+++ src/core-api.client.ts
@@ -7,6 +7,7 @@ import {  AccountAPIService } from './core/account/account-api.service';
 import { TranslationApiService } from './core/translations/translations-api.service';
 import { ServiceManagementAPIService } from './core/service-management/service-management.service';
 import { DataServiceAPIService } from './core/data-service/data-service.service';
+import { RulesAPIService } from './core/rules/rules-api.service';
 
 
 export class CoreAPIClient {
@@ -16,6 +17,7 @@ export class CoreAPIClient {
 
   private _accountApi: AccountAPIService;
   private _translationApi: TranslationApiService;
+  private _rulesApi: RulesAPIService;
 
   private _dataServiceAPI: DataServiceAPIService
 
@@ -87,6 +89,7 @@ export class CoreAPIClient {
     this._accountApi = new AccountAPIService(this._config, this._auth);
     this._dataServiceAPI = new DataServiceAPIService(this._config, _http, this._auth);
     this._translationApi = new TranslationApiService(this._config, _http, this._auth);
+    this._rulesApi = new RulesAPIService(this._config, _http, this._auth);
     this._serviceManagementApi = new ServiceManagementAPIService(this._config, _http, this._auth);
   }
 
@@ -135,6 +138,15 @@ export class CoreAPIClient {
    */
   public get translationAPI(): TranslationApiService {
     return this._translationApi;
+  }
+
+  /**
+   * Provides access to the Rules API for managing business rules.
+   * 
+   * @returns {RulesAPIService} Rules API instance for business rule operations.
+   */
+  public get rulesAPI(): RulesAPIService {
+    return this._rulesApi;
   }
 
   /**